+++ settings.gradle
@@ -162,6 +162,7 @@ include "support:cas-server-support-google-analytics"
 include "support:cas-server-support-grouper"
 include "support:cas-server-support-grouper-core"
 include "support:cas-server-support-gua"
+include "support:cas-server-support-gua-webflow"
 include "support:cas-server-support-hazelcast"
 include "support:cas-server-support-hazelcast-core"
 include "support:cas-server-support-hazelcast-discovery-aws"
