+++ support/cas-server-support-gua-webflow/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.apereo.cas.gua.config.GraphicalUserAuthenticationWebflowConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.apereo.cas.config.GraphicalUserAuthenticationWebflowConfiguration
